Refactor routes and auth, remove legacy pages and tests
- Remove legacy routes: /login, /me, /about, /wishlist/*
- Remove related test files and wishlist item component
- Update dashboard breadcrumb and navigation paths
- Refactor auth logic for redirects and session checks
- Add credentials: 'include' to auth client fetch options
- Simplify home page with landing card and dashboard/login/signup links
- Update vitest config to include test file globs
- Enable crossSubDomainCookies only in production in auth package
