+++ apps/web/src/lib/auth-client.ts
@@ -4,4 +4,7 @@ import { env } from '../env';
 
 export const authClient = createAuthClient({
 	baseURL: env.VITE_AUTH_URL,
+	fetchOptions: {
+		credentials: 'include',
+	},
 });
